+++ src/models/PanelSettings.ts
@@ -65,6 +65,10 @@ export interface Field {
   fieldGroup?: string | string[];
   dataType?: string | string[];
   taxonomyLimit?: number;
+
+  // Date fields
+  isPublishDate?: boolean;
+  isModifiedDate?: boolean;
 }
 
 export interface DateInfo {